requirements.txt
Flask
PyMuPDF
Pillow